requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-dotenv>=1.0.1
httpx>=0.26.0
jinja2>=3.1.3
aiofiles>=23.2.1
python-multipart>=0.0.7
typing-extensions>=4.9.0
types-requests>=2.31.0.20240125
types-fastapi-v0.103.1
aiohttp>=3.9.3
python-dateutil>=2.8.2
pytest>=7.4.4
pytest-asyncio>=0.23.5
mypy>=1.8.0
htmx.org>=1.9.10 